Added hyperstructure representation
Any hyperstructure is identify by a matrix. Each entry in this matrix is an integer, whose binary representation corresponds to a subset of H.
We defined a tag function that takes the matrix as input, converts its entries into binary and concatenates all of them. This results into a vector of 1s and 0s.
The corresponding integer to this vector is the representation of the matrix (and therefore of the hyperstructure).
The inverse function is also available: this means that we can generate an hyperstructure by and integer.