Updated to U1024
We are starting to use uint U1024 to represent tags. In this way we can represent hyperstrictures of order up to 10.
Everything is working as before, but some improvment and modification is still in progress.
In particular, enumeration algorithms, that use parallel iterator, are now slower, because U1024 doesn't implement Iter and ParallelIter;
however, we can make a parallel iteration using .par_bridge(), but it is less efficient.